+++ src/components/pages/identity-overview/identity-overview.tsx
@@ -2,7 +2,10 @@ import React from "react";
 import styled from "styled-components";
 
 import { Identity, IdentityTypes } from "@lib/@types";
-import { AwaitingValidationBadge } from "@src/components/awaiting-validation-badge/awaiting-validation-badge";
+import {
+  AwaitingValidationBadge,
+  PrimaryBadge,
+} from "@src/components/badges/badges";
 import { Box } from "@src/components/box/box";
 import { Button, ButtonVariant } from "@src/components/buttons/buttons";
 import { ConfirmationBox } from "@src/components/confirmation-box/confirmation-box";
@@ -10,7 +13,6 @@ import { DeleteConfirmationBoxContent } from "@src/components/confirmation-box/d
 import { PrimaryConfirmationBoxContent } from "@src/components/confirmation-box/primary-confirmation-box-content";
 import { FakeButton } from "@src/components/fake-button/fake-button";
 import { NeutralLink } from "@src/components/neutral-link/neutral-link";
-import { PrimaryBadge } from "@src/components/primary-badge/primary-badge";
 import { getIdentityType } from "@src/utils/get-identity-type";
 
 type IdentityOverviewProps = {
